# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Slot Game Backend API
  version: 0.1.0
  description: |
    老虎机游戏后端 API 文档（基于当前代码实现整理）。
    注意：部分响应结构含有嵌套对象/动态字段，此处以通用对象占位。
servers:
  - url: http://localhost:8080
    description: Local Dev
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StartRequest:
      type: object
      required: [bet_amount]
      properties:
        bet_amount:
          type: integer
          minimum: 100
    StartResponse:
      type: object
      properties:
        session_id:
          type: string
        balance:
          type: integer
        message:
          type: string
    SpinRequest:
      type: object
      required: [session_id]
      properties:
        session_id:
          type: string
    SpinResponse:
      type: object
      properties:
        result:
          type: object
          description: 旋转结果（结构随主题而定）
        balance:
          type: integer
        state:
          type: string
        total_bet:
          type: integer
        total_win:
          type: integer
    BatchSpinRequest:
      type: object
      required: [session_id, spin_count]
      properties:
        session_id:
          type: string
        spin_count:
          type: integer
          minimum: 1
          maximum: 100
        auto_stop:
          type: boolean
        stop_on_big_win:
          type: boolean
        big_win_amount:
          type: integer
    BatchSpinResponse:
      type: object
      properties:
        session_id:
          type: string
        spin_results:
          type: array
          items:
            type: object
        total_spins:
          type: integer
        total_bet:
          type: integer
        total_win:
          type: integer
        state:
          type: string
        stopped_early:
          type: boolean
        stop_reason:
          type: string
        final_balance:
          type: integer
    SettleRequest:
      type: object
      required: [session_id]
      properties:
        session_id:
          type: string
    SettleResponse:
      type: object
      properties:
        total_bet:
          type: integer
        total_win:
          type: integer
        balance:
          type: integer
        rtp:
          type: number
          format: float
        message:
          type: string
    HistoryResponse:
      type: object
      properties:
        records:
          type: array
          items:
            type: object
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
    SessionInfoResponse:
      type: object
      properties:
        session_id:
          type: string
        state:
          type: string
        spin_count:
          type: integer
        total_bet:
          type: integer
        total_win:
          type: integer
        rtp:
          type: number
        duration:
          type: number
        last_result:
          type: object
    UserGameStats:
      type: object
      properties:
        user_id:
          type: integer
        total_games:
          type: integer
        total_bet:
          type: integer
        total_win:
          type: integer
        biggest_win:
          type: integer
        last_played:
          type: string
          format: date-time
        rtp:
          type: number
paths:
...
  /api/v1/slot/start:
    post:
      tags: [Slot]
      summary: 开始游戏
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartResponse'
  /api/v1/slot/spin:
    post:
      tags: [Slot]
      summary: 单次转动
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpinRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpinResponse'
  /api/v1/slot/batch-spin:
    post:
      tags: [Slot]
      summary: 批量转动
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchSpinRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchSpinResponse'
  /api/v1/slot/settle:
    post:
      tags: [Slot]
      summary: 结算
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettleRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettleResponse'
  /api/v1/slot/history:
    get:
      tags: [Slot]
      summary: 历史记录
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: page_size
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
  /api/v1/slot/session/{id}:
    get:
      tags: [Slot]
      summary: 会话信息
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionInfoResponse'
  /api/v1/slot/stats:
    get:
      tags: [Slot]
      summary: 用户统计
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGameStats'